requirements.txt
Pillow~=9.5.0
opencv-python~=4.8.0.76
pynput~=1.7.6
pytesseract~=0.3.10